More robust way to find ivy resolution file.
We use a range of versions of sbt (including our own forks), and have
struggled with this piece of code. This tries to be a bit more robust
... instead of assuming behavior about versions, it just looks
everywhere the file might be, and of the files that actually exist,
takes the most recent version. This is sort flawed in theory, useful in
practice.